+++ src/main/java/com/websystique/springmvc/service/CourseService.java
@@ -8,7 +8,7 @@ import java.util.List;
  * Created by kevin on 2016/11/28.
  */
 public interface CourseService {
-    Course findById(Integer id);
+    Course findById(Long id);
 
     Course findByClassGradeIdAndSubjectName(Integer classGradeId, String subjectsStr);
